Query parser refactoring: delete Parser Q-S
