+++ src/main/java/org/elasticsearch/index/query/SpanFirstQueryBuilder.java
@@ -19,7 +19,13 @@
 
 package org.elasticsearch.index.query;
 
+import org.apache.lucene.search.Query;
+import org.apache.lucene.search.spans.SpanFirstQuery;
+import org.apache.lucene.search.spans.SpanQuery;
+import org.elasticsearch.common.Strings;
+import org.elasticsearch.common.inject.Inject;
 import org.elasticsearch.common.xcontent.XContentBuilder;
+import org.elasticsearch.common.xcontent.XContentParser;
 
 import java.io.IOException;
 
@@ -35,6 +41,14 @@ public class SpanFirstQueryBuilder extends BaseQueryBuilder implements SpanQuery
     private float boost = -1;
 
     private String queryName;
+    
+    public static final String NAME = "span_first";
+
+    @Inject
+    public SpanFirstQueryBuilder() {
+        this.matchBuilder = null;
+        this.end = 0;
+    }
 
     public SpanFirstQueryBuilder(SpanQueryBuilder matchBuilder, int end) {
         this.matchBuilder = matchBuilder;
@@ -57,7 +71,7 @@ public class SpanFirstQueryBuilder extends BaseQueryBuilder implements SpanQuery
 
     @Override
     protected void doXContent(XContentBuilder builder, Params params) throws IOException {
-        builder.startObject(SpanFirstQueryParser.NAME);
+        builder.startObject(SpanFirstQueryBuilder.NAME);
         builder.field("match");
         matchBuilder.toXContent(builder, params);
         builder.field("end", end);
@@ -68,5 +82,62 @@ public class SpanFirstQueryBuilder extends BaseQueryBuilder implements SpanQuery
             builder.field("name", queryName);
         }
         builder.endObject();
+    }
+    
+    @Override
+    public String[] names() {
+        return new String[]{NAME, Strings.toCamelCase(NAME)};
+    }
+
+    @Override
+    public Query parse(QueryParseContext parseContext) throws IOException, QueryParsingException {
+        XContentParser parser = parseContext.parser();
+
+        float boost = 1.0f;
+
+        SpanQuery match = null;
+        int end = -1;
+        String queryName = null;
+
+        String currentFieldName = null;
+        XContentParser.Token token;
+        while ((token = parser.nextToken()) != XContentParser.Token.END_OBJECT) {
+            if (token == XContentParser.Token.FIELD_NAME) {
+                currentFieldName = parser.currentName();
+            } else if (token == XContentParser.Token.START_OBJECT) {
+                if ("match".equals(currentFieldName)) {
+                    Query query = parseContext.parseInnerQuery();
+                    if (!(query instanceof SpanQuery)) {
+                        throw new QueryParsingException(parseContext.index(), "spanFirst [match] must be of type span query");
+                    }
+                    match = (SpanQuery) query;
+                } else {
+                    throw new QueryParsingException(parseContext.index(), "[span_first] query does not support [" + currentFieldName + "]");
+                }
+            } else {
+                if ("boost".equals(currentFieldName)) {
+                    boost = parser.floatValue();
+                } else if ("end".equals(currentFieldName)) {
+                    end = parser.intValue();
+                } else if ("_name".equals(currentFieldName)) {
+                    queryName = parser.text();
+                } else {
+                    throw new QueryParsingException(parseContext.index(), "[span_first] query does not support [" + currentFieldName + "]");
+                }
+            }
+        }
+        if (match == null) {
+            throw new QueryParsingException(parseContext.index(), "spanFirst must have [match] span query clause");
+        }
+        if (end == -1) {
+            throw new QueryParsingException(parseContext.index(), "spanFirst must have [end] set for it");
+        }
+
+        SpanFirstQuery query = new SpanFirstQuery(match, end);
+        query.setBoost(boost);
+        if (queryName != null) {
+            parseContext.addNamedQuery(queryName, query);
+        }
+        return query;
     }
 }
