+++ src/main/java/org/elasticsearch/index/query/SpanNearQueryBuilder.java
@@ -19,11 +19,20 @@
 
 package org.elasticsearch.index.query;
 
+import org.apache.lucene.search.Query;
+import org.apache.lucene.search.spans.SpanNearQuery;
+import org.apache.lucene.search.spans.SpanQuery;
 import org.elasticsearch.ElasticsearchIllegalArgumentException;
+import org.elasticsearch.common.Strings;
+import org.elasticsearch.common.inject.Inject;
 import org.elasticsearch.common.xcontent.XContentBuilder;
+import org.elasticsearch.common.xcontent.XContentParser;
 
 import java.io.IOException;
 import java.util.ArrayList;
+import java.util.List;
+
+import static com.google.common.collect.Lists.newArrayList;
 
 /**
  *
@@ -41,6 +50,12 @@ public class SpanNearQueryBuilder extends BaseQueryBuilder implements SpanQueryB
     private float boost = -1;
 
     private String queryName;
+    
+    public static final String NAME = "span_near";
+
+    @Inject
+    public SpanNearQueryBuilder() {
+    }
 
     public SpanNearQueryBuilder clause(SpanQueryBuilder clause) {
         clauses.add(clause);
@@ -84,7 +99,7 @@ public class SpanNearQueryBuilder extends BaseQueryBuilder implements SpanQueryB
         if (slop == null) {
             throw new ElasticsearchIllegalArgumentException("Must set the slop when building a spanNear query");
         }
-        builder.startObject(SpanNearQueryParser.NAME);
+        builder.startObject(SpanNearQueryBuilder.NAME);
         builder.startArray("clauses");
         for (SpanQueryBuilder clause : clauses) {
             clause.toXContent(builder, params);
@@ -104,5 +119,72 @@ public class SpanNearQueryBuilder extends BaseQueryBuilder implements SpanQueryB
             builder.field("_name", queryName);
         }
         builder.endObject();
+    }
+    
+    @Override
+    public String[] names() {
+        return new String[]{NAME, Strings.toCamelCase(NAME)};
+    }
+
+    @Override
+    public Query parse(QueryParseContext parseContext) throws IOException, QueryParsingException {
+        XContentParser parser = parseContext.parser();
+
+        float boost = 1.0f;
+        Integer slop = null;
+        boolean inOrder = true;
+        boolean collectPayloads = true;
+        String queryName = null;
+
+        List<SpanQuery> clauses = newArrayList();
+
+        String currentFieldName = null;
+        XContentParser.Token token;
+        while ((token = parser.nextToken()) != XContentParser.Token.END_OBJECT) {
+            if (token == XContentParser.Token.FIELD_NAME) {
+                currentFieldName = parser.currentName();
+            } else if (token == XContentParser.Token.START_ARRAY) {
+                if ("clauses".equals(currentFieldName)) {
+                    while ((token = parser.nextToken()) != XContentParser.Token.END_ARRAY) {
+                        Query query = parseContext.parseInnerQuery();
+                        if (!(query instanceof SpanQuery)) {
+                            throw new QueryParsingException(parseContext.index(), "spanNear [clauses] must be of type span query");
+                        }
+                        clauses.add((SpanQuery) query);
+                    }
+                } else {
+                    throw new QueryParsingException(parseContext.index(), "[span_near] query does not support [" + currentFieldName + "]");
+                }
+            } else if (token.isValue()) {
+                if ("in_order".equals(currentFieldName) || "inOrder".equals(currentFieldName)) {
+                    inOrder = parser.booleanValue();
+                } else if ("collect_payloads".equals(currentFieldName) || "collectPayloads".equals(currentFieldName)) {
+                    collectPayloads = parser.booleanValue();
+                } else if ("slop".equals(currentFieldName)) {
+                    slop = Integer.valueOf(parser.intValue());
+                } else if ("boost".equals(currentFieldName)) {
+                    boost = parser.floatValue();
+                } else if ("_name".equals(currentFieldName)) {
+                    queryName = parser.text();
+                } else {
+                    throw new QueryParsingException(parseContext.index(), "[span_near] query does not support [" + currentFieldName + "]");
+                }
+            } else {
+                throw new QueryParsingException(parseContext.index(), "[span_near] query does not support [" + currentFieldName + "]");
+            }
+        }
+        if (clauses.isEmpty()) {
+            throw new QueryParsingException(parseContext.index(), "span_near must include [clauses]");
+        }
+        if (slop == null) {
+            throw new QueryParsingException(parseContext.index(), "span_near must include [slop]");
+        }
+
+        SpanNearQuery query = new SpanNearQuery(clauses.toArray(new SpanQuery[clauses.size()]), slop.intValue(), inOrder, collectPayloads);
+        query.setBoost(boost);
+        if (queryName != null) {
+            parseContext.addNamedQuery(queryName, query);
+        }
+        return query;
     }
 }
