+++ src/main/java/org/elasticsearch/indices/query/IndicesQueriesModule.java
@@ -75,33 +75,33 @@ public class IndicesQueriesModule extends AbstractModule {
         qpBinders.addBinding().to(DisMaxQueryBuilder.class).asEagerSingleton();
         qpBinders.addBinding().to(IdsQueryBuilder.class).asEagerSingleton();
         qpBinders.addBinding().to(MatchAllQueryBuilder.class).asEagerSingleton();
-        qpBinders.addBinding().to(QueryStringQueryParser.class).asEagerSingleton();
+        qpBinders.addBinding().to(QueryStringQueryBuilder.class).asEagerSingleton();
         qpBinders.addBinding().to(BoostingQueryBuilder.class).asEagerSingleton();
         qpBinders.addBinding().to(BoolQueryBuilder.class).asEagerSingleton();
         qpBinders.addBinding().to(TermQueryParser.class).asEagerSingleton();
         qpBinders.addBinding().to(TermsQueryParser.class).asEagerSingleton();
         qpBinders.addBinding().to(FuzzyQueryBuilder.class).asEagerSingleton();
-        qpBinders.addBinding().to(RegexpQueryParser.class).asEagerSingleton();
-        qpBinders.addBinding().to(RangeQueryParser.class).asEagerSingleton();
+        qpBinders.addBinding().to(RegexpQueryBuilder.class).asEagerSingleton();
+        qpBinders.addBinding().to(RangeQueryBuilder.class).asEagerSingleton();
         qpBinders.addBinding().to(PrefixQueryBuilder.class).asEagerSingleton();
         qpBinders.addBinding().to(WildcardQueryParser.class).asEagerSingleton();
         qpBinders.addBinding().to(FilteredQueryBuilder.class).asEagerSingleton();
         qpBinders.addBinding().to(ConstantScoreQueryBuilder.class).asEagerSingleton();
-        qpBinders.addBinding().to(SpanTermQueryParser.class).asEagerSingleton();
-        qpBinders.addBinding().to(SpanNotQueryParser.class).asEagerSingleton();
+        qpBinders.addBinding().to(SpanTermQueryBuilder.class).asEagerSingleton();
+        qpBinders.addBinding().to(SpanNotQueryBuilder.class).asEagerSingleton();
         qpBinders.addBinding().to(FieldMaskingSpanQueryBuilder.class).asEagerSingleton();
-        qpBinders.addBinding().to(SpanFirstQueryParser.class).asEagerSingleton();
-        qpBinders.addBinding().to(SpanNearQueryParser.class).asEagerSingleton();
-        qpBinders.addBinding().to(SpanOrQueryParser.class).asEagerSingleton();
+        qpBinders.addBinding().to(SpanFirstQueryBuilder.class).asEagerSingleton();
+        qpBinders.addBinding().to(SpanNearQueryBuilder.class).asEagerSingleton();
+        qpBinders.addBinding().to(SpanOrQueryBuilder.class).asEagerSingleton();
         qpBinders.addBinding().to(MoreLikeThisQueryBuilder.class).asEagerSingleton();
         qpBinders.addBinding().to(FuzzyLikeThisQueryBuilder.class).asEagerSingleton();
         qpBinders.addBinding().to(FuzzyLikeThisFieldQueryBuilder.class).asEagerSingleton();
         qpBinders.addBinding().to(WrapperQueryParser.class).asEagerSingleton();
         qpBinders.addBinding().to(IndicesQueryBuilder.class).asEagerSingleton();
         qpBinders.addBinding().to(CommonTermsQueryBuilder.class).asEagerSingleton();
-        qpBinders.addBinding().to(SpanMultiTermQueryParser.class).asEagerSingleton();
+        qpBinders.addBinding().to(SpanMultiTermQueryBuilder.class).asEagerSingleton();
         qpBinders.addBinding().to(FunctionScoreQueryParser.class).asEagerSingleton();
-        qpBinders.addBinding().to(SimpleQueryStringParser.class).asEagerSingleton();
+        qpBinders.addBinding().to(SimpleQueryStringBuilder.class).asEagerSingleton();
         qpBinders.addBinding().to(TemplateQueryParser.class).asEagerSingleton();
 
         if (ShapesAvailability.JTS_AVAILABLE) {
